Fix make gen command for Go workspace
The 'go generate ./...' command was failing in the Go workspace setup
because the root directory is not a module. This changes the gen target
to run go generate in each workspace module directory individually.

Fixes the GoReleaser action failure where 'make gen' was failing with:
pattern ./...: directory prefix . does not contain modules listed in go.work